refactored project repo class that deals with props genericly
